--- src/modules/mahjong/Hand/version/hongkong/HandNames.ts
@@ -1,21 +1,0 @@
-export enum HKHandNames {
-  PING_WU = 'ping wu (all consecutives)',
-  DEOI_DEOI_WU = 'deoi deoi wu (all triplets)',
-  WAN_JAT_SIK = 'wan jat sik (semi-purity)',
-  CING_JAT_SIK = 'cing jat sik (purity)',
-  ZI_JAT_SIK = 'zi jat sik (all honor tiles)',
-  SIU_SAAM_JYUN = 'siu saam jyun (small dragons)',
-  DAAI_SAAM_JYUN = 'daai saam jyun (large dragons)',
-  SIU_SEI_HEI = 'siu sei hei (small winds)',
-  DAAI_SEI_HEI = 'daai sei hei (large winds)',
-  SAP_SAAM_JIU = 'sap saam jiu (thirteen orphans)',
-}
-
-export enum AdditionalPoints {
-  DRAGON = 'dragon',
-  WIND = 'wind',
-  CONCEALED = 'concealed',
-  SELF_DRAW = 'self draw',
-  FLOWER = 'flower',
-  SEASON = 'season',
-}
