+++ src/modules/mahjong/enums/BonusTileEnums.ts
