+++ src/modules/mahjong/enums/HandPointsEnums.ts
@@ -1,0 +1,22 @@
+export enum HKHandPoints {
+  ALL_CONSECUTIVE = 1,
+  ALL_TRIPLETS = 3,
+  SEMI_PURITY = 3,
+  PURITY = 7,
+  ALL_HONORS = 10,
+  SMALL_DRAGONS = 5,
+  LARGE_DRAGONS = 8,
+  SMALL_WINDS = 10,
+  LARGE_WINDS = 13,
+  THIRTEEN_ORPHANS = 13,
+  ALL_KONGS = 13,
+}
+
+export enum AdditionalPoints {
+  DRAGON = 1,
+  WIND = 1,
+  CONCEALED = 1,
+  SELF_DRAW = 1,
+  FLOWER = 1,
+  SEASON = 1,
+}
