+++ src/modules/mahjong/enums/HonorTileEnums.ts
