+++ src/modules/mahjong/enums/SimpleTileEnums.ts
