+++ src/modules/mahjong/enums/VersionsEnum.ts
